requirements.txt
django==1.9.2
python-dateutil==2.5.3
python-frontmatter==0.3.1
google-api-python-client
markdown==2.6.7
requests==2.9.1
# google-api-python-client==1.3.2